Replace u32 with usize
For non-STV methods, this is very slightly slower for few voters,
and very slightly faster for more voters, on a 64-bit machine,
because usize is essentially u64. The differences are not
noticeable though, and probably not statistically significant.
The benefits is better readability and maintenance,
which outweighs the negligible performance hit.

For STV, this is very slightly faster in all 6 situations
tested in the STV config.